Change Owner & Borrower to Interface
